Remove recovery_partition suite Move the one "useful" partition test into test_zzzrecovery
